Move ParameterFile definition to parameter_descriptions.py
