+++ launch_ros/launch_ros/descriptions/__init__.py
@@ -15,8 +15,8 @@
 """descriptions Module."""
 
 from .composable_node import ComposableNode
-from .parameter_file import ParameterFile
 from ..parameter_descriptions import Parameter
+from ..parameter_descriptions import ParameterFile
 from ..parameter_descriptions import ParameterValue
